fix: removed about and changed stuff in login register negative
