+++ webapp/e2e/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-    testMatch: ["**/playing_full_game_pos*.steps.js","**/about.steps.js", "**/login_positive.steps.js","**/about_positive_logged_user_seeing_about_screen.steps.js"
+    testMatch: ["**/playing_full_game_pos*.steps.js","**/login_register_negative_bad_email_format.steps.js", "**/login_positive.steps.js","**/about_positive_logged_user_seeing_about_screen.steps.js"
     ,"**/about_positive_non_logged_user_seeing_about_screen.steps.js"],
     moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
     testTimeout: 30000
